Update main.js
Update main.js, change validating updatingArrays methods. Put while loop and prompt method to keep asking until we get result not 1, so we have collide with some kind of trouble that game will stop.

Made switch statment instead of if else to move character and then we validate this moves